# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Nsangusa API
  version: 1.0.0
  description: Static language-neutral contract for implemented Spring MVC routes.
    Spring Security supplies the explicitly marked login and logout operations.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
- url: https://api.example.invalid
  description: Placeholder; configure per environment
...
security:
- sessionCookie: []
- basicAuth: []
paths:
...
  "/api/v1/admin/image-generations/{generationId}/content":
    get:
      tags: [Media]
      operationId: previewEditorialImage
      summary: Preview an image candidate without making it public
      x-required-roles: [EDITOR, ADMINISTRATOR]
      parameters:
      - "$ref": "#/components/parameters/GenerationId"
      - "$ref": "#/components/parameters/ImageVariant"
      responses:
        '200':
          description: Private optimized preview; no-store
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/svg+xml:
              schema:
                type: string
                format: binary
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          description: Image candidate or stored variant not found
        default:
          "$ref": "#/components/responses/UnexpectedError"
...
  "/api/v1/admin/articles/{articleId}/images":
    get:
      tags:
      - Media
      operationId: listArticleImageGenerations
      summary: List image generations for an article
      x-required-roles:
      - EDITOR
      - ADMINISTRATOR
      parameters:
      - "$ref": "#/components/parameters/ArticleId"
      responses:
        '200':
          description: Image generations
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/ImageGeneration"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '429':
          "$ref": "#/components/responses/TooManyRequests"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
        default:
          "$ref": "#/components/responses/UnexpectedError"
  "/api/v1/admin/articles/{articleId}/images/regenerate":
    post:
      tags:
      - Media
      operationId: regenerateArticleImage
      summary: Request a new generated image for an article
      x-required-roles:
      - EDITOR
      - ADMINISTRATOR
      parameters:
      - "$ref": "#/components/parameters/IdempotencyKey"
      - "$ref": "#/components/parameters/ArticleId"
      - "$ref": "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ImageRegenerationRequest"
      responses:
        '202':
          description: Image-generation event accepted
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/EventResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '409':
          "$ref": "#/components/responses/Conflict"
        '429':
          "$ref": "#/components/responses/TooManyRequests"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
        default:
          "$ref": "#/components/responses/UnexpectedError"
...
  "/api/v1/admin/image-generations/{generationId}/approve":
    post:
      tags:
      - Media
      operationId: approveImageGeneration
      summary: Approve an image generation
      x-required-roles:
      - EDITOR
      - ADMINISTRATOR
      parameters:
      - "$ref": "#/components/parameters/IdempotencyKey"
      - "$ref": "#/components/parameters/GenerationId"
      - "$ref": "#/components/parameters/CsrfToken"
      responses:
        '204':
          description: Operation completed
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '409':
          "$ref": "#/components/responses/Conflict"
        '429':
          "$ref": "#/components/responses/TooManyRequests"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
        default:
          "$ref": "#/components/responses/UnexpectedError"
  "/api/v1/admin/media-assets/{assetId}":
    delete:
      tags:
      - Media
      operationId: deleteMediaAsset
      summary: Delete a media asset from object storage and soft-delete its record
      x-required-roles:
      - EDITOR
      - ADMINISTRATOR
      parameters:
      - "$ref": "#/components/parameters/IdempotencyKey"
      - "$ref": "#/components/parameters/AssetId"
      - "$ref": "#/components/parameters/CsrfToken"
      responses:
        '204':
          description: Operation completed
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '409':
          "$ref": "#/components/responses/Conflict"
        '429':
          "$ref": "#/components/responses/TooManyRequests"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
        default:
          "$ref": "#/components/responses/UnexpectedError"
...
  /api/v1/admin/users:
    get:
      tags:
      - Identity
      operationId: listIdentityUsers
      summary: Search and page user accounts for administration
      x-required-roles:
      - ADMINISTRATOR
      description: >
        Requires administrator authorization and rechecks current database roles. Searches display name and
        email case-insensitively with literal substring matching; percent and underscore are not wildcard
        operators. Omitting status includes all non-deleted accounts. Ordering is createdAt descending, then id
        ascending.
      parameters:
      - name: q
        in: query
        required: false
        schema:
          type: string
          maxLength: 100
      - name: role
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/Phase4IdentityRole'
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - active
          - unverified
          - disabled
          - deleted
        description: active means enabled and verified; unverified means enabled but not verified.
      - name: page
        in: query
        schema:
          type: integer
          minimum: 0
          maximum: 100000
          default: 0
      - $ref: '#/components/parameters/WorkspaceSize'
      responses:
        '200':
          description: Bounded user page.
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Phase4IdentityUserPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /api/v1/admin/users/{id}:
    get:
      tags:
      - Identity
      operationId: getIdentityUser
      summary: Read an account's current administrative detail and version
      x-required-roles:
      - ADMINISTRATOR
      parameters:
      - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Current user detail.
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Phase4IdentityAdminUser'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Account not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /api/v1/admin/users/{id}/roles:
    put:
      tags:
      - Identity
      operationId: replaceIdentityUserRoles
      summary: Confirm a versioned replacement of another account's allowed roles
      x-required-roles:
      - ADMINISTRATOR
      description: >
        Requires a durable actor-scoped request key, the loaded user version and an exact copy of the selected
        account's email as confirmation. Replaces the complete role set. Administrators cannot change their own
        roles. The target must be enabled, verified and non-deleted; the last active administrator cannot be
        removed. Actor identity comes from the authenticated principal. Authorization is rechecked before
        receipt lookup, including retries. Identical actor/key/target/version/payload retries return 204 without
        a second mutation or audit. Reusing the same key for a different payload returns 409. Role order is
        canonicalized before fingerprinting. Successful changes are audited, revoke target sessions after commit
        and change a persistent security stamp. Locally managed roles are not restored from external-provider
        claims on later login.
      parameters:
      - $ref: '#/components/parameters/Id'
      - $ref: '#/components/parameters/CsrfToken'
      - name: Idempotency-Key
        in: header
        required: true
        description: Unlike legacy editorial operations, this operation requires a request key.
        schema:
          type: string
          minLength: 8
          maxLength: 200
          pattern: ^[!-~]{8,200}$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Phase4IdentityRoleChange'
      responses:
        '204':
          description: Roles replaced, or the same successful command replayed.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Account not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        default:
          $ref: '#/components/responses/UnexpectedError'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
      description: Redis-backed opaque Spring Session cookie
    basicAuth:
      type: http
      scheme: basic
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Actor-scoped retry key. Reusing a key with another target or payload returns 409. Success receipts and domain/outbox state commit together; absent keys retain legacy execution.
      schema:
        type: string
        minLength: 8
        maxLength: 200
        pattern: '^[!-~]{8,200}$'
...
    WorkspaceSize:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    ImageVariant:
      name: variant
      in: query
      schema:
        type: string
        enum: [hero, thumbnail, social]
        default: hero
    Id:
      name: id
      in: path
      required: true
      schema:
        "$ref": "#/components/schemas/Identifier"
    ArticleId:
      name: articleId
      in: path
      required: true
      schema:
        "$ref": "#/components/schemas/Identifier"
...
    GenerationId:
      name: generationId
      in: path
      required: true
      schema:
        "$ref": "#/components/schemas/Identifier"
    AssetId:
      name: assetId
      in: path
      required: true
      schema:
        "$ref": "#/components/schemas/Identifier"
...
    CsrfToken:
      name: X-XSRF-TOKEN
      in: header
      required: true
      description: Spring Security CSRF token matching the XSRF-TOKEN cookie
      schema:
        type: string
        minLength: 1
...
  responses:
    BadRequest:
      description: Invalid path, query, header, or request-body input
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
...
    Unauthorized:
      description: Authentication is required or failed
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
    Forbidden:
      description: Authentication succeeded but authorization or CSRF validation failed
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
    Conflict:
      description: The requested transition conflicts with current resource state
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
    TooManyRequests:
      description: Per-client request rate exceeded
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
      headers:
        Retry-After:
          schema:
            type: integer
          description: Retry delay in seconds
    ServiceUnavailable:
      description: Rate limiting could not reach its Redis dependency
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
    UnexpectedError:
      description: Unexpected API failure
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
  schemas:
    Identifier:
      type: string
      format: uuid
...
    EventResponse:
      type: object
      additionalProperties: false
      required:
      - eventId
      properties:
        eventId:
          "$ref": "#/components/schemas/Identifier"
...
    Problem:
      type: object
      additionalProperties: true
      required:
      - title
      - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
...
    ImageGeneration:
      type: object
      additionalProperties: false
      required:
      - id
      - articleId
      - prompt
      - altText
      - objectKey
      - provider
      - model
      - safetyStatus
      - createdAt
      - approvedAt
      properties:
        id:
          "$ref": "#/components/schemas/Identifier"
        articleId:
          "$ref": "#/components/schemas/Identifier"
        prompt:
          type: string
        altText:
          type: string
        objectKey:
          type:
          - string
          - 'null'
        provider:
          type:
          - string
          - 'null'
        model:
          type:
          - string
          - 'null'
        safetyStatus:
          type:
          - string
          - 'null'
        createdAt:
          type: string
          format: date-time
        approvedAt:
          type:
          - string
          - 'null'
          format: date-time
...
    ImageRegenerationRequest:
      type: object
      additionalProperties: false
      required:
      - prompt
      - altText
      properties:
        prompt:
          type: string
          minLength: 1
          maxLength: 4000
        altText:
          type: string
          minLength: 1
          maxLength: 500
...
    Phase4IdentityRole:
      type: string
      enum:
      - READER
      - MODERATOR
      - EDITOR
      - ADMINISTRATOR
    Phase4IdentityVersion:
      type: integer
      format: int64
      minimum: 0
...
    Phase4IdentityAdminUser:
      type: object
      additionalProperties: false
      required:
      - id
      - email
      - displayName
      - roles
      - emailVerified
      - enabled
      - deletedAt
      - createdAt
      - lastLoginAt
      - version
      - rolesManagedLocally
      properties:
        id:
          $ref: '#/components/schemas/Identifier'
        email:
          type: string
          format: email
          maxLength: 320
        displayName:
          type: string
          maxLength: 100
        roles:
          type: array
          uniqueItems: true
          maxItems: 4
          items:
            $ref: '#/components/schemas/Phase4IdentityRole'
        emailVerified:
          type: boolean
        enabled:
          type: boolean
        deletedAt:
          type:
          - string
          - 'null'
          format: date-time
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type:
          - string
          - 'null'
          format: date-time
        version:
          $ref: '#/components/schemas/Phase4IdentityVersion'
        rolesManagedLocally:
          type: boolean
          description: External-provider role claims cannot add roles once local administration has taken ownership.
    Phase4IdentityUserPage:
      type: object
      additionalProperties: false
      required:
      - items
      - page
      - size
      - total
      properties:
        items:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/Phase4IdentityAdminUser'
        page:
          type: integer
          minimum: 0
          maximum: 100000
        size:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          format: int64
          minimum: 0
    Phase4IdentityRoleChange:
      type: object
      additionalProperties: false
      required:
      - roles
      - expectedVersion
      - confirmation
      properties:
        roles:
          type: array
          minItems: 1
          maxItems: 4
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Phase4IdentityRole'
        expectedVersion:
          $ref: '#/components/schemas/Phase4IdentityVersion'
        confirmation:
          type: string
          minLength: 1
          maxLength: 320
          pattern: \S
          description: Must exactly equal the selected account's current email.
...